JAVA-2505: Move MongoDriverInformation from com.mongodb.client package to com.mongodb package.
   This breaks binary compatibility, but without this change it's impossible to modularize the core driver,
   as the com.mongodb.client package would be split between two modules, which is not allowed.

   A mitigating factor is that this class is only intended for use by wrapping drivers like our own Scala
   and Reactive Streams drivers, and so is unlikely to be referenced directly by downstream users besides
   MongoDB itself.
